+++ ghost/members-api/package.json
@@ -35,7 +35,6 @@
     "@tryghost/logging": "2.4.21",
     "@tryghost/magic-link": "0.0.0",
     "@tryghost/member-events": "0.0.0",
-    "@tryghost/members-payments": "0.0.0",
     "@tryghost/nql": "0.12.7",
     "@tryghost/tpl": "0.1.32",
     "@tryghost/validator": "0.2.14",
