--- ghost/payments/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@tryghost/members-payments",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/domain-events": "0.0.0",
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/members-offers": "0.0.0",
-    "@tryghost/tiers": "0.0.0"
-  }
-}
